feat(notifications): functions moved to utility files
