+++ src/layouts/MultiDhoLayout.vue
@@ -1,9 +1,9 @@
 <script>
 import { mapActions, mapGetters, mapMutations } from 'vuex'
 import BrowserIpfs from '~/ipfs/browser-ipfs.js'
-import I18n from '~/utils/i18n'
-import { date } from 'quasar'
 import gql from 'graphql-tag'
+import { timeago } from '~/utils/TimeUtils'
+import { parsedNotification } from '~/utils/notifications-utils'
 
 const NOTIFICATIONS_QUERY = `
   queryNotification {
@@ -100,6 +100,8 @@ export default {
 
   data () {
     return {
+      timeago,
+      parsedNotification,
       profile: {
         username: null,
         avatar: null,
@@ -378,90 +380,6 @@ export default {
         this.$router.push({ name: 'proposals' })
       } catch (error) {
 
-      }
-    },
-
-    timeago (createdDate) {
-      const TODAY = new Date().toUTCString()
-      const created = new Date(createdDate).toUTCString()
-      const second = date.getDateDiff(created, TODAY, 'seconds')
-      const minute = date.getDateDiff(created, TODAY, 'minutes')
-      const hour = date.getDateDiff(created, TODAY, 'hours')
-      const day = date.getDateDiff(created, TODAY, 'days')
-      const month = date.getDateDiff(created, TODAY, 'months')
-      const year = date.getDateDiff(created, TODAY, 'years')
-      if (year > 0) {
-        if (month < 12) {
-          return `${month} month${month > 1 ? 's' : ''} ago`
-        } else {
-          return `${year} year${year > 1 ? 's' : ''} ago`
-        }
-      }
-      if (month > 0) return `${month} month${month > 1 ? 's' : ''} ago`
-      if (day > 0) return `${day} day${day > 1 ? 's' : ''} ago`
-      if (hour > 0) return `${hour} hour${hour > 1 ? 's' : ''} ago`
-      if (minute > 0) return `${minute} minute${minute > 1 ? 's' : ''} ago`
-      if (second > 0) return `${second} second${second > 1 ? 's' : ''} ago`
-
-      return ''
-    },
-
-    parsedNotification (notification) {
-      let icon = null
-      let title = null
-      let description = null
-      const createdDate = this.timeago(notification.time)
-      switch (notification.event.name) {
-        case ('notification.test'):
-          icon = require('~/assets/icons/notifications/newcomment.png')
-          title = notification.content
-          description = notification.content
-          break
-        case ('proposals.comment'):
-          icon = require('~/assets/icons/notifications/newcomment.png')
-          title = I18n.t('notifications.newComment')
-          description = I18n.t('notifications.hasJustLeftAComment', { accountname: notification.user })
-          break
-        case ('proposals.voting_expiry'):
-          icon = require('~/assets/icons/notifications/voting-expire.png')
-          title = I18n.t('notifications.proposalVotingExpire')
-          description = I18n.t('notifications.proposalIsExpiring', { accountname: notification.user, days: notification.content?.days })
-          break
-        case ('proposals.passed'):
-          icon = require('~/assets/icons/notifications/proposal-passed.png')
-          title = I18n.t('notifications.proposalPassed')
-          description = I18n.t('notifications.proposalHasPassed', { accountname: notification.user })
-          break
-        case ('proposals.rejected'):
-          icon = require('~/assets/icons/notifications/proposal-rejected.png')
-          title = I18n.t('notifications.proposalRejected')
-          description = I18n.t('notifications.proposalHasntPassed', { accountname: notification.user })
-          break
-        case ('system.claimable_period'):
-          icon = require('~/assets/icons/notifications/claimable-period.png')
-          title = I18n.t('notifications.claimablePeriod')
-          description = I18n.t('notifications.youHaveClaimablePeriod', { value: notification.content?.periods })
-          break
-        case ('assignment.extension_expiry'):
-          icon = require('~/assets/icons/notifications/extend-assignment.png')
-          title = I18n.t('notifications.extendYourAssignment')
-          description = I18n.t('notifications.youStillHave', { days: notification.content?.days })
-          break
-        case ('assignment.approved'):
-          icon = require('~/assets/icons/notifications/assignment-approved.png')
-          title = I18n.t('notifications.assignmentApproved')
-          description = I18n.t('notifications.yourAssignmentHasBeenApproved')
-          break
-        case ('assignment.rejected'):
-          icon = require('~/assets/icons/notifications/assignment-rejected.png')
-          title = I18n.t('notifications.assignmentRejected')
-          description = I18n.t('notifications.yourAssignmentHasntBeenApproved')
-      }
-      return {
-        icon: icon,
-        title: title,
-        description: description,
-        createdDate: createdDate
       }
     },
     initNotifications () {
